requirements.txt
streamlit==1.32.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
plotly==5.18.0
pandas==2.2.0